moved comments to posts
