+++ stod-backend/stod/groups/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-from django.contrib.auth.models import User
 
 
 class Group(models.Model):
@@ -9,16 +8,3 @@ class Group(models.Model):
 
     def __str__(self):
         return self.name
-
-
-class Comment(models.Model):
-    name = models.CharField(max_length=140)
-    email = models.EmailField()
-    body = models.TextField()
-    created_on = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        ordering = ['created_on']
-
-    def __str__(self):
-        return 'Comment {} by {}'.format(self.body, self.name)
