+++ stod-backend/stod/posts/admin.py
@@ -1,3 +1,9 @@
 from django.contrib import admin
+from .models import Comment
 
-# Register your models here.
+@admin.register(Comment)
+class CommentAdmin(admin.ModelAdmin):
+    list_display = ('name', 'body', 'created_on')
+    list_filter = ('created_on',)
+    search_fields = ('name', 'email', 'body')
+
