+++ stod-backend/stod/posts/models.py
@@ -12,3 +12,15 @@ class Post (models.Model):
     def __str__(self):
         """A string representation of the model."""
         return self.title
+
+class Comment(models.Model):
+    name = models.CharField(max_length=140)
+    email = models.EmailField()
+    body = models.TextField()
+    created_on = models.DateTimeField(auto_now_add=True)
+
+    class Meta:
+        ordering = ['created_on']
+
+    def __str__(self):
+        return 'Comment {} by {}'.format(self.body, self.name)
